+++ subscribers/market_depth_test.go
@@ -46,22 +46,21 @@ func buildOrder(id string, side types.Side, orderType types.OrderType, price uin
 }
 
 type OrderEventWithVegaTime struct {
-	events.Event
-	order    *vega.Order
+	events.Order
 	vegaTime time.Time
 }
 
-func (oe OrderEventWithVegaTime) VegaTime() time.Time {
+func (oe *OrderEventWithVegaTime) VegaTime() time.Time {
 	return oe.vegaTime
 }
 
-func (oe OrderEventWithVegaTime) Order() *vega.Order {
-	return oe.order
+func (oe *OrderEventWithVegaTime) GetOrder() *vega.Order {
+	return oe.Order.Order()
 }
 
-func newOrderEvent(ctx context.Context, o *types.Order) OrderEventWithVegaTime {
+func newOrderEvent(ctx context.Context, o *types.Order) *OrderEventWithVegaTime {
 	oe := events.NewOrderEvent(ctx, o)
-	return OrderEventWithVegaTime{oe, oe.Order(), time.Now()}
+	return &OrderEventWithVegaTime{*oe, time.Now()}
 }
 
 func TestBuyPriceLevels(t *testing.T) {